refactor: appeal service
